Factor Model

Factor model object from features2 data set using polychoric correlations from rho2. This is the output of the psych::fa function. This model has 12 features and 6 factors. See vignette("making-gov-scores", package = "governance") for a detailed description of how and why this model was chosen.

Most notable internal objects include

loadings

Loadings from the factor analysis using polychoric correlations of features2.

scores

Scores using the PEARSON correlations. Note, we do not use the scores generated from the psych::fa function as they always use the Pearson correlation, even if the factor loadings were caclulated using the polychoric correlations.

Usage

model6

Format

An object of class psych (inherits from fa) of length 51.